+++ ck2/cmind/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.7.6"
+__version__ = "0.7.6.1"
 
 from cmind.core import access
 from cmind.core import error